Using unsigned int to avoid overflow issues

+++ BinarySearchRec.cpp
@@ -11,7 +11,7 @@ using namespace std;
 class BinarySearchRec : public BasePred {
 private:
     // x is the value we are looking for, p is the closest so far
-    int x, p;
+    unsigned int x, p;
 
 public:
     /**
@@ -20,19 +20,19 @@ public:
      * @param x
      * @return
      */
-    int pred(int x) {
+    unsigned int pred(unsigned int x) {
         this->x = x;
-        return binarySearch(0, (int) ((this->array->size()) - 1));
+        return binarySearch(0, ((this->array->size()) - 1));
     }
 
-    int binarySearch(int l, int r) {
+    unsigned int binarySearch(unsigned int l, unsigned int r) {
         if(l > r){
             // we are done, return the closest matching
             return this->p;
         }
 
         // position of the middle element
-        int m = floor((l+r)/2);
+        unsigned int m = floor((l+r)/2);
 
         if(this->array->at(m) < x) {
             // save this as closest
