+++ LinearScanPred.cpp
@@ -14,12 +14,12 @@ public:
      * @param x
      * @return
      */
-    int pred(int x) {
-        int max = 1;
+    unsigned int pred(unsigned int x) {
+        unsigned int max = 1;
 
-        int size = (int) this->array->size();
+        unsigned int size = this->array->size();
 
-        for (int i = 0; i < size; ++i) {
+        for (unsigned int i = 0; i < size; ++i) {
             if (this->array->at(i) > max && this->array->at(i) <= x) {
                 max = this->array->at(i);
             }
